Renewed AI Agent and Trainer, testing on laptop now
